+++ pricing/src/epilot/models/shared/order.py
@@ -47,6 +47,7 @@ class Products:
 @dataclasses.dataclass
 class Order:
     r"""The order entity"""
+    UNSET='__SPEAKEASY_UNSET__'
     additional_properties: Optional[Dict[str, Any]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'exclude': lambda f: f is None }})
     created_at: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('_created_at'), 'exclude': lambda f: f is None }})
     r"""The order creation date"""