+++ pricing/src/epilot/models/shared/price.py
@@ -87,6 +87,7 @@ class PriceSchemas1(str, Enum):
 @dataclasses.dataclass
 class Price:
     r"""The price entity schema for simple pricing"""
+    UNSET='__SPEAKEASY_UNSET__'
     additional_properties: Optional[Dict[str, Any]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'exclude': lambda f: f is None }})
     created_at: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('_created_at'), 'exclude': lambda f: f is None }})
     r"""The price creation date"""